Agents/GitHubSearchAgent/requirements.txt
requests>=2.31.0
PyGithub>=1.59.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
python-dotenv>=1.0.0
dataclasses-json>=0.6.0
pydantic>=2.0.0
openai>=1.0.0
tiktoken>=0.5.0
tenacity>=8.2.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
gitpython>=3.1.0
pygments>=2.16.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-java>=0.20.0
